Move classes related to MagpieServer to magpiebridge package
