+++ src/main/java/Main.java
@@ -13,6 +13,9 @@ import gobpie.GobPieConfiguration;
 import gobpie.GobPieException;
 import magpiebridge.core.MagpieServer;
 import magpiebridge.core.ServerAnalysis;
+import magpiebridge.GoblintLanguageExtensionHandler;
+import magpiebridge.GoblintMagpieServer;
+import magpiebridge.GoblintServerConfiguration;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.eclipse.lsp4j.MessageParams;
